Keep spectrum strength math centralized
Move the existing amplitude-to-dB conversion helper out of ui_spectrum_controller.ts so the Live spectrum inspector can use strength metrics without tripping the UI static guard.
