+++ apps/ui/src/app/runtime/ui_spectrum_controller.ts
@@ -3,18 +3,15 @@ import type uPlot from "uplot";
 import {
   SPECTRUM_DB_MAX,
   SPECTRUM_DB_MIN,
-  SPECTRUM_DB_REFERENCE_AMP_G,
-  SPECTRUM_MIN_RENDER_AMP_G,
   SPECTRUM_TWEEN_DURATION_MS,
 } from "../../config";
 import { escapeHtml } from "../../format";
-import { SpectrumChart } from "../../spectrum";
+import { convertSpectrumAmplitudesToDbInPlace, SpectrumChart } from "../../spectrum";
 import { chartSeriesPalette, orderBandFills } from "../../theme";
 import { areHeavyFramesCompatible, interpolateHeavyFrame, type SpectrumHeavyFrame } from "../spectrum_animation";
 import type { UiDomElements } from "../ui_dom_registry";
 import type { AppState, ChartBand } from "../ui_app_state";
 
-const SPECTRUM_LOG10_REF = Math.log10(SPECTRUM_DB_REFERENCE_AMP_G);
 const FREQ_MATCH_EPSILON = 1e-6;
 
 /** Check if two freq arrays match up to `len` elements. */
@@ -49,18 +46,6 @@ function interpolateToTarget(
     out[i] = f1 <= f0 ? v0 : v0 + ((v1 - v0) * ((freq - f0) / (f1 - f0)));
   }
   return out;
-}
-
-/** Convert amplitude array to dB scale in-place. */
-function convertToDbInPlace(values: number[]): void {
-  for (let i = 0; i < values.length; i++) {
-    const amplitude = values[i];
-    const safe = Number.isFinite(amplitude) && amplitude > 0
-      ? Math.max(amplitude, SPECTRUM_MIN_RENDER_AMP_G)
-      : SPECTRUM_MIN_RENDER_AMP_G;
-    const db = 20 * (Math.log10(safe) - SPECTRUM_LOG10_REF);
-    values[i] = Math.max(SPECTRUM_DB_MIN, Math.min(SPECTRUM_DB_MAX, db));
-  }
 }
 
 const bandKeyPresentation: Record<string, { color: string; labelKey: string }> = {
@@ -381,7 +366,7 @@ export class UiSpectrumController {
         // Source array — must copy before mutating.
         blended = blended.slice();
       }
-      convertToDbInPlace(blended);
+      convertSpectrumAmplitudesToDbInPlace(blended);
 
       entries.push({
         id: client.id,
